Remove committed virtualenv files which should be refs #26
Remove the virtualenv directory and files named env recently committed. All
developers should use virtualenvs, but it is better if these are not committed
to git. virtualenvs directories are large and will differ on each developer's
machine depending on the version of python they have installed.

The .gitignore file should have caught the env as an ignored name, but I think
the web upload interface doesn't consult .gitigore.

On Linux the standard way to create a virtualenv is as follows:

python3 -m venv venv && source venv/bin/activate

And directory venv/ should not be committed.